x/omni/types/message_observe_vote_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
	"omni/testutil/sample"
)

func TestMsgObserveVote_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgObserveVote
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgObserveVote{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgObserveVote{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
